+++ R/theme_tidybde.R
@@ -38,7 +38,7 @@
 #'   names(series_TC_pivot) <- c("x", "y")
 #'
 #'   ggplot(series_TC_pivot, aes(x = x, y = y)) +
-#'     geom_line(linewidth = 0.8, color = bde_palettes(n = 1)) +
+#'     geom_line(linewidth = 0.8, color = bde_tidy_palettes(n = 1)) +
 #'     labs(
 #'       title = "Title",
 #'       subtitle = "Some metric",