+++ NEWS.md
@@ -1,12 +1,13 @@
-# tidyBdE (development version)
+# tidyBdE 0.3.5
 
 Mostly changes on the color functions:
 
 -   `bde_vivid_pal()` and `bde_rose_pal()` have been superseded. Use the new
-    function `bde_palettes()` instead.
--   `scale_color_bde_d()` and friends leverage now on `bde_palettes()`, and
+    function `bde_tidy_palettes()` instead.
+-   `scale_color_bde_d()` and friends leverage now on `bde_tidy_palettes()`, and
     these functions gain two new arguments: `alpha` y `rev`.
 -   Update and review documentation.
+-   New palette named `bde_qual_pal`.
 
 # tidyBdE 0.3.4